Combining the formData reducer and actions list into a slice. We can continue building our reducer with extraReducers(), so there was never a need to split this into more verbose code.
