+++ src/altinn-app-frontend/src/features/form/dynamics/conditionalRendering/conditionalRenderingSagas.ts
@@ -2,8 +2,8 @@ import type { SagaIterator } from 'redux-saga';
 import { all, call, put, select, take, takeLatest } from 'redux-saga/effects';
 import type { IRuntimeState, IValidations, IUiConfig } from 'src/types';
 import { runConditionalRenderingRules } from '../../../../utils/conditionalRendering';
-import FormDataActions from '../../data/formDataActions';
-import type { IFormData } from '../../data/formDataReducer';
+import { FormDataActions } from '../../data/formDataSlice';
+import type { IFormData } from '../../data';
 import { FormLayoutActions } from '../../layout/formLayoutSlice';
 import { updateValidations } from '../../validation/validationSlice';
 import { FormDynamicsActions } from 'src/features/form/dynamics/formDynamicsSlice';
