+++ src/altinn-app-frontend/src/selectors/getErrors.test.ts
@@ -3,7 +3,7 @@ import { getInitialStateMock } from '../../__mocks__/initialStateMock';
 import type { IFormDynamicState } from 'src/features/form/dynamics';
 import type { IFormRuleState } from 'src/features/form/rules';
 import type { ITextResourcesState } from 'src/shared/resources/textResources';
-import type { IFormDataState } from '../features/form/data/formDataReducer';
+import type { IFormDataState } from '../features/form/data';
 
 import { makeGetHasErrorsSelector } from './getErrors';
