+++ src/altinn-app-frontend/src/shared/resources/attachments/map/mapAttachmentsSagas.ts
@@ -7,8 +7,8 @@ import { mapAttachmentListToAttachments } from 'src/utils/attachment';
 import { AttachmentActions } from 'src/shared/resources/attachments/attachmentSlice';
 import type { IApplicationMetadata } from '../../applicationMetadata';
 import { getCurrentTaskData } from 'src/utils/appMetadata';
-import type { IFormData } from 'src/features/form/data/formDataReducer';
-import FormDataActions from 'src/features/form/data/formDataActions';
+import type { IFormData } from 'src/features/form/data';
+import { FormDataActions } from 'src/features/form/data/formDataSlice';
 import type { ILayouts } from 'src/features/form/layout';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
 import { ApplicationMetadataActions } from 'src/shared/resources/applicationMetadata/applicationMetadataSlice';
