+++ src/altinn-app-frontend/src/shared/resources/attachments/upload/uploadAttachmentSagas.ts
@@ -8,7 +8,7 @@ import { getFileUploadComponentValidations } from '../../../../utils/formCompone
 import type { IRuntimeState } from '../../../../types';
 import { post } from '../../../../utils/networking';
 import { fileUploadUrl } from '../../../../utils/appUrlHelper';
-import FormDataActions from 'src/features/form/data/formDataActions';
+import { FormDataActions } from 'src/features/form/data/formDataSlice';
 import type { ILanguage } from 'altinn-shared/types';
 import { AttachmentActions } from 'src/shared/resources/attachments/attachmentSlice';
 import type { PayloadAction } from '@reduxjs/toolkit';
