+++ src/altinn-app-frontend/src/shared/resources/options/fetch/fetchOptionsSagas.ts
@@ -22,10 +22,10 @@ import { get } from 'altinn-shared/utils';
 import { getOptionsUrl } from 'src/utils/appUrlHelper';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
 import { OptionsActions } from '../optionsSlice';
-import FormDataActions from 'src/features/form/data/formDataActions';
+import { FormDataActions } from 'src/features/form/data/formDataSlice';
 import type { IUpdateFormDataFulfilled } from 'src/features/form/data/formDataTypes';
 import type { PayloadAction } from '@reduxjs/toolkit';
-import type { IFormData } from 'src/features/form/data/formDataReducer';
+import type { IFormData } from 'src/features/form/data';
 import { getOptionLookupKey } from 'src/utils/options';
 import { appLanguageStateSelector } from 'src/selectors/appLanguageStateSelector';
 import { LanguageActions } from 'src/shared/resources/language/languageSlice';
