+++ src/altinn-app-frontend/src/shared/resources/queue/dataTask/dataTaskQueueSagas.test.ts
@@ -2,7 +2,7 @@ import { expectSaga } from 'redux-saga-test-plan';
 
 import { startInitialDataTaskQueueSaga } from './dataTaskQueueSagas';
 import { startInitialDataTaskQueueFulfilled } from '../queueSlice';
-import FormDataActions from 'src/features/form/data/formDataActions';
+import { FormDataActions } from 'src/features/form/data/formDataSlice';
 import { fetchJsonSchema } from 'src/features/form/datamodel/datamodelSlice';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
