+++ src/altinn-app-frontend/src/shared/resources/queue/dataTask/dataTaskQueueSagas.ts
@@ -1,7 +1,7 @@
 import type { SagaIterator } from 'redux-saga';
 import { put, takeEvery } from 'redux-saga/effects';
 import { fetchJsonSchema } from 'src/features/form/datamodel/datamodelSlice';
-import FormDataActions from '../../../../features/form/data/formDataActions';
+import { FormDataActions } from '../../../../features/form/data/formDataSlice';
 import { FormLayoutActions } from '../../../../features/form/layout/formLayoutSlice';
 import {
   startInitialDataTaskQueue,
