+++ src/altinn-app-frontend/src/shared/resources/textResources/replace/replaceTextResourceSagas.test.ts
@@ -1,6 +1,6 @@
 import { testSaga } from 'redux-saga-test-plan';
 import { take } from 'redux-saga/effects';
-import FormDataActions from 'src/features/form/data/formDataActions';
+import { FormDataActions } from 'src/features/form/data/formDataSlice';
 import { FormLayoutActions } from 'src/features/form/layout/formLayoutSlice';
 import {
   replaceTextResourcesSaga,
